Reflect 2019-6-11 errata to the reference implementation
The errata fixed week number displayed by ~V to align ISO definition,
but the reference implementation wasn't updated accordingly.
This patch updates the reference implementation; the behavior matches
'%V' formatter of C99.  Note that it differs from date-week-number in
the way more than the first day of the week offset; the partial week
at the beginning of the year can belong to the last week of the previous
year, or vice versa.

A new file srfi-19-gauche.scm is to load srfi-19.scm into Gauche.
Ideally we want RnRS portable adapter, but that needs more changes.